refactor(app): change structure of the project
