--- .vscode/launch.json
@@ -1,15 +1,0 @@
-{
-    // Use IntelliSense to learn about possible attributes.
-    // Hover to view descriptions of existing attributes.
-    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-    "version": "0.2.0",
-    "configurations": [
-        {
-            "type": "pwa-chrome",
-            "request": "launch",
-            "name": "Launch Chrome against localhost",
-            "url": "file:///C:/Users/Eng/Documents/LifeLink/ThingsBoard_API_Painel/index.html",
-            "webRoot": "${workspaceFolder}"
-        }
-    ]
-}
