+++ app/src/generate-entities/devices.ts
