+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
     "compilerOptions": {
-        "outDir": "dist/js",
+        "outDir": "app/dist/js",
         "target":  "ES6",
         "noEmitOnError": true,
         "noImplicitAny": true,
